+++ grails-app/controllers/io/xh/hoist/admin/ClientAdminController.groovy
@@ -9,12 +9,12 @@ package io.xh.hoist.admin
 
 import io.xh.hoist.BaseController
 import static io.xh.hoist.json.JSONParser.parseArray
-import io.xh.hoist.security.Access
+import io.xh.hoist.security.AccessRequiresRole
 
 import static io.xh.hoist.util.ClusterUtils.runOnAllInstancesAsJson
 import static io.xh.hoist.util.ClusterUtils.runOnInstanceAsJson
 
-@Access(['HOIST_ADMIN_READER'])
+@AccessRequiresRole('HOIST_ADMIN_READER')
 class ClientAdminController extends BaseController {
 
     def webSocketService
@@ -26,7 +26,7 @@ class ClientAdminController extends BaseController {
         renderJSON(ret)
     }
 
-    @Access(['HOIST_ADMIN'])
+    @AccessRequiresRole('HOIST_ADMIN')
     def pushToClient(String channelKey, String topic, String message, String instance) {
         def ret = runOnInstanceAsJson(webSocketService.&pushToChannel, instance, [channelKey, topic, message])
         renderClusterJSON(ret)