+++ grails-app/controllers/io/xh/hoist/admin/cluster/ConnectionPoolMonitorAdminController.groovy
@@ -7,10 +7,11 @@
 package io.xh.hoist.admin.cluster
 
 import io.xh.hoist.BaseController
-import io.xh.hoist.security.Access
+import io.xh.hoist.security.AccessRequiresRole
+
 import static io.xh.hoist.util.ClusterUtils.runOnInstanceAsJson
 
-@Access(['HOIST_ADMIN_READER'])
+@AccessRequiresRole('HOIST_ADMIN_READER')
 class ConnectionPoolMonitorAdminController extends BaseController {
 
     def connectionPoolMonitoringService
@@ -20,13 +21,13 @@ class ConnectionPoolMonitorAdminController extends BaseController {
         renderClusterJSON(ret)
     }
 
-    @Access(['HOIST_ADMIN'])
+    @AccessRequiresRole('HOIST_ADMIN')
     def takeSnapshot(String instance) {
         def ret = runOnInstanceAsJson(connectionPoolMonitoringService.&takeSnapshot, instance)
         renderClusterJSON(ret)
     }
 
-    @Access(['HOIST_ADMIN'])
+    @AccessRequiresRole('HOIST_ADMIN')
     def resetStats(String instance) {
         def ret = runOnInstanceAsJson(connectionPoolMonitoringService.&resetStats, instance)
         renderClusterJSON(ret)