+++ src/main/groovy/io/xh/hoist/security/Access.groovy
@@ -13,11 +13,14 @@ import java.lang.annotation.Retention
 import java.lang.annotation.RetentionPolicy
 import java.lang.annotation.Target
 
+
 /**
  * Controller annotation to list roles required to execute any action.
  * Current user must have ALL roles listed to access.
- * Equivalent to @RequiresEvery.
+ *
+ * @deprecated - use @AccessRequiresRole or @AccessRequiresAllRoles instead.
  */
+@Deprecated
 @Inherited
 @Target([ElementType.METHOD, ElementType.TYPE])
 @Retention(RetentionPolicy.RUNTIME)