+++ src/main/groovy/io/xh/hoist/security/AccessRequiresAllRoles.groovy
@@ -12,12 +12,11 @@ import java.lang.annotation.*
 /**
  * Controller annotation to list roles required to execute any action.
  * Current user must have ALL roles listed to access.
- * Equivalent to @Access.
  */
 @Inherited
 @Target([ElementType.METHOD, ElementType.TYPE])
 @Retention(RetentionPolicy.RUNTIME)
-@interface RequiresEvery {
-    /** Array of role names required. */
-    String[] value() default []
+@interface AccessRequiresAllRoles {
+    /** Array of role names, all of which are required for access. */
+    String[] value()
 }